Integration of Erant's Wii DVD-Loading via libdi and Wii Zip ROM support (Use: 7za a -tzip -mx=9 rom.zip rom.z64). GC ROM-Cache needs a re-work to add in Zip support, then some of the nastyness of my quick-fix hacks to get this compiling for both GC and Wii will be gone!

+++ fileBrowser/fileBrowser.h
@@ -5,6 +5,8 @@
 #ifndef FILE_BROWSER_H
 #define FILE_BROWSER_H
 
+#include <stdint.h>
+
 #define FILE_BROWSER_MAX_PATH_LEN 128
 
 #define FILE_BROWSER_ATTR_DIR     0x10
@@ -18,7 +20,7 @@
 
 typedef struct {
 	char         name[FILE_BROWSER_MAX_PATH_LEN];
-	unsigned int discoffset; // Only necessary for DVD
+	uint64_t discoffset; // Only necessary for DVD
 	unsigned int offset; // Keep track of our offset in the file
 	unsigned int size;
 	unsigned int attr;
